Changed recommended logic
