+++ src/controllers/phoneCardData.ts
@@ -16,6 +16,13 @@ export const getById = async(req: Request, res: Response) => {
   res.send(requiredPhoneCardData);
 };
 
+export const getRecomendedPhones = async(req: Request, res: Response) => {
+  const { id } = req.params;
+  const phones = await phoneCardDataService.getRecomendedPhones(id);
+
+  res.send(phones);
+};
+
 export const getImagesById = async(req: Request, res: Response) => {
   const { id } = req.params;
   const requiredImages = await phoneCardDataService.getImagesById(String(id));
