+++ src/routers/phoneCardData.ts
@@ -8,4 +8,5 @@ export const router = express.Router();
 
 router.get('/', phoneCardDataController.getAll);
 router.get('/:id', phoneCardDataController.getById);
+router.get('/:id/recomended', phoneCardDataController.getRecomendedPhones);
 router.get('/:id/images', phoneCardDataController.getImagesById);
