+++ src/routers/phones.ts
@@ -10,6 +10,5 @@ router.get('/pagination', phonesController.getByPagination);
 router.get('/newest', phonesController.getNewestPhones);
 router.get('/discount', phonesController.getByDiscount);
 
+router.get('/:imagePath', phonesController.getPhoneImage);
 router.get('/:id', phonesController.getById);
-router.get('/:id/image', phonesController.getPhoneImage);
-router.get('/:id/recomended', phonesController.getRecomendedPhones);
