+++ src/services/phones.ts
@@ -14,10 +14,8 @@ export const getById = async(id: number) => {
   return phone;
 };
 
-export const getPhoneImage = async(id: number) => {
-  const phone = await getById(id);
-
-  const path = `public/${phone?.image}`;
+export const getImageByPath = async(imagePath: string) => {
+  const path = `public/${imagePath}`;
 
   if (path) {
     const image = fs.readFileSync(path);
@@ -89,25 +87,6 @@ export const getByPagination = async(
   };
 
   return info;
-};
-
-export const getRecomendedPhones = async(id: string) => {
-  const phoneArray = await getAll();
-  const chosenPhone = await Phone.findByPk(id);
-
-  if (!chosenPhone) {
-    return [];
-  }
-
-  const recommendedPhones = phoneArray.filter((phone) => {
-    return (
-      phone.price <= chosenPhone.price + 50
-      && phone.price >= chosenPhone.price - 50
-      && phone.id !== chosenPhone.id
-    );
-  });
-
-  return recommendedPhones.slice(0, 8);
 };
 
 export const getNewestPhones = async() => {
